Moved the deletion of unused packages to the dependency sub-package.
